custom time allocation mutator uses the new config group
git-svn-id: file:///srv/local/data/cleanRepo/trunk@20746 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/thibautd/src/main/java/playground/thibautd/jointtrips/replanning/modules/jointtimeallocationmutator/JointTimeAllocationMutatorAlgorithm.java
@@ -45,7 +45,7 @@ public class JointTimeAllocationMutatorAlgorithm implements PlanAlgorithm {
 	public JointTimeAllocationMutatorAlgorithm(
 			final Random random,
 			final StageActivityTypes routerStageTypes,
-			final int mutationRange) {
+			final double mutationRange) {
 		this.random = random;
 		CompositeStageActivityTypes blackList = new CompositeStageActivityTypes();
 		blackList.addActivityTypes( routerStageTypes );
